requirements.txt
# Core
flask>=2.0.0
flask-cors>=3.0.0

# LangChain + Google Gemini
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-google-genai>=0.1.0
langchain-chroma>=0.1.0

# PostgreSQL (ogms)
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Data processing
pandas>=1.5.0

# Embeddings and vector storage
chromadb>=0.4.0

# Pydantic for data validation
pydantic>=2.0.0

# Google AI
google-generativeai>=0.3.0

# Additional utilities
python-dotenv>=0.19.0
